[enh] Upgrade to OwnCloud 8.0.3

+++ sources/lib/private/cache/fileglobal.php
@@ -52,7 +52,7 @@ class FileGlobal {
 	public function hasKey($key) {
 		$key = $this->fixKey($key);
 		$cache_dir = self::getCacheDir();
-		if ($cache_dir && is_file($cache_dir.$key)) {
+		if ($cache_dir && is_file($cache_dir.$key) && is_readable($cache_dir.$key)) {
 			$mtime = filemtime($cache_dir.$key);
 			if ($mtime < time()) {
 				unlink($cache_dir.$key);
@@ -81,31 +81,6 @@ class FileGlobal {
 				while (($file = readdir($dh)) !== false) {
 					if($file!='.' and $file!='..' and ($prefix==='' || strpos($file, $prefix) === 0)) {
 						unlink($cache_dir.$file);
-					}
-				}
-			}
-		}
-	}
-
-	static public function gc() {
-		$appConfig = \OC::$server->getAppConfig();
-		$last_run = $appConfig->getValue('core', 'global_cache_gc_lastrun', 0);
-		$now = time();
-		if (($now - $last_run) < 300) {
-			// only do cleanup every 5 minutes
-			return;
-		}
-		$appConfig->setValue('core', 'global_cache_gc_lastrun', $now);
-		$cache_dir = self::getCacheDir();
-		if($cache_dir and is_dir($cache_dir)) {
-			$dh=opendir($cache_dir);
-			if(is_resource($dh)) {
-				while (($file = readdir($dh)) !== false) {
-					if($file!='.' and $file!='..') {
-						$mtime = filemtime($cache_dir.$file);
-						if ($mtime < $now) {
-							unlink($cache_dir.$file);
-						}
 					}
 				}
 			}
